+++ prefix-hex/src/data.rs
@@ -19,12 +19,6 @@ impl FromHexPrefixed for Box<[u8]> {
     }
 }
 
-impl ToHexPrefixed for Vec<u8> {
-    fn to_hex_prefixed(self) -> String {
-        format!("0x{}", hex::encode(self))
-    }
-}
-
 impl<const N: usize> FromHexPrefixed for [u8; N]
 where
     Self: hex::FromHex,
@@ -61,7 +55,7 @@ where
     }
 }
 
-macro_rules! impl_for_as_ref_type {
+macro_rules! impl_to_hex {
     ($type:ty) => {
         impl ToHexPrefixed for $type {
             fn to_hex_prefixed(self) -> String {
@@ -71,6 +65,8 @@ macro_rules! impl_for_as_ref_type {
     };
 }
 
-impl_for_as_ref_type!(Box<[u8]>);
-impl_for_as_ref_type!(&Box<[u8]>);
-impl_for_as_ref_type!(&[u8]);
+impl_to_hex!(Box<[u8]>);
+impl_to_hex!(&Box<[u8]>);
+impl_to_hex!(&[u8]);
+impl_to_hex!(Vec<u8>);
+impl_to_hex!(&Vec<u8>);