+++ prefix-hex/CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `FromHexPrefixed` impl for boxed slice;
+- `ToHexPrefixed` impl for Vec references;
 
 ## 0.7.0 - 2023-03-14